Make service type constants of type ServiceType Set all model ids for accessory, service and characteristic to invalid by default. New id is only assigend when id is invalid which allows us to remove and add accessories and not have their ids updated

+++ model/accessory.go
@@ -5,7 +5,7 @@ type Accessory interface {
     Compareable
     
     // Returns the accessory id
-    GetId()int
+    GetId() int64
     
     // Returns the services which represent the accessory 
     GetServices()[]Service
